teague_scenarios/requirements.txt
pyyaml
docker
PyGithub
python-on-whales
#locust # https://docs.locust.io/en/stable/configuration.html